requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.8
numpy==1.18.0
pandas==0.25.3
python-dateutil==2.8.1
python-http-client==3.2.1
pytz==2019.3
requests==2.22.0
schedule==0.6.0
sendgrid==6.1.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.25.7
Flask==1.1.1
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
python-decouple==3.4